downgrade "expected" warning message to info message
when there is an attempt to insert data into the columns cache, there is
currently no check if the columns cache actually has a size > 0. if the
columns cache is configured with a size of 0, then all insert attempts
will fail and produce the warning

"Failed to allocate memory for buffered column id x, name: x of size x"

the warning on its own looks scary, but it is totally expected in case
the columns cache is configured with a size of 0 (which is the default
value in ArangoDB).

this PR changes the message text so it is more clear what could have
caused the problem, and also downgrades the log level from WARN to INFO.

the message text is now extended with the following text:

"This can happen if no columns cache was configured or the column data
exceeds the columns cache size."